Update: Category tages added to post and home
Each post will not have a category or categories assigned to them. This
category is displayed in the header of the post and is a link the
archive by category page specific to the selected category.

The archive by category also displays all the categories as links to
their respective page as well as an "all" tag which will show a
page with all categories listed with a list of each categories articles.

The landing page is updated to display all the tags. The tags are links
to their respective archive by category page.

All blog pages besides landing include the home_button displayed on the
upper right of the screen (opposite of main navigation) and will return
user to blog landing page.