style (navigation) updated navigation bar
Recreate the navigation bar and color palette of the site. The
navigation bar now appears solid across the top of the page with
toggle menu on the right and svg logo on the left. HTML updated
to accommodate this change. All styles not utilized commented out
for further update to each section.

Remove navigation_style.css as it was no longer needed.

.SCSS files have been updated to new color palette

script.js updated including new class names for navigation elements.