feat (blog) complete blog page
Complete the blog page of the site. This includes the 3 most
recent posts appearing on the landing page with a button to move
to the blog page (/blog/). The blog page consists on the next 3 most
recent posts with buttons to move through the pages of the blog.

jekyll-paginate-v2 added to the gem file and comfig file with appropriate
configuration.

style.css updated to include all blog, blog post, and button stylings.