feat (about page) complete the about me page
Add an about the blog and about me sections to the about page. Includes
placeholders to Medium profile link and publication to be updated
at a later date.

All posting have new front matter (image, image_by, and image_by_link)
to specify the image to be used for the post, the name of the unsplash
photographer and the link to there profile. After downloading an image
copy the supplied link and extract required information. All images
display the name and link to the profile and unsplash.com.

style.css updated to support the styling of the about page. Similar to
blog but always vertical, sections never go side by side.

Removed zone identifier files for downloaded images.