feat (projects) complete projects page
Complete projects page, set up in a similar fashion as the blog page.
Create the _projects directory to house the project markdown file
which store the project attributes to be displayed on the project
card. New attributes of img_by and img_by_link were added.

Add all images for blog posts and projects with associated metadata

Updates to the config.yml file to support the project collections

Updates to style.css to support styling of the projects page.