feat (post updates) update markdown to html
Posts have been updated to accommodate the ability to write the
posts as jupyter notebooks. Update and include appropriate style
sheets for styling of input/output cells of the notebook. The jupyter
notebooks then can be exported (nbconvert) to html --no-prompt and
all but the resulting <body> can be removed, creating a post. The
notebooks have been added to the _notebooks directory.

Update url in config to ethanguyant.com.

Include Medium and publication links in footer.html

Project directory removed and is now included as a general blog post
with the category Project. In main navigation Project link brings
user to the Project category page.

landing.html, bytag.html, bycategory.html, blog.html all updated to
show post Category and Tag prior to the post description.